Refactored rules/style test cases
This separates positive and negative test cases for rules. (#527)
