+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/KT.kt
@@ -48,7 +48,8 @@ enum class Case(val file: String) {
 	UnreachableCode("/cases/UnreachableCode.kt"),
 	UnnecessaryAbstractClassPositive("/cases/UnnecessaryAbstractClassPositive.kt"),
 	UnnecessaryAbstractClassNegative("/cases/UnnecessaryAbstractClassNegative.kt"),
-	UtilityClasses("/cases/UtilityClasses.kt"),
+	UtilityClassesPositive("/cases/UtilityClassesPositive.kt"),
+	UtilityClassesNegative("/cases/UtilityClassesNegative.kt"),
 	SerializablePositive("/cases/SerializablePositive.kt"),
 	SerializableNegative("/cases/SerializableNegative.kt"),
 	SuppressedElements("/SuppressedByElementAnnotation.kt"),
@@ -59,7 +60,8 @@ enum class Case(val file: String) {
 	SwallowedExceptionPositive("/cases/SwallowedExceptionPositive.kt"),
 	TooManyFunctions("/cases/TooManyFunctions.kt"),
 	TooManyFunctionsTopLevel("/cases/TooManyFunctionsTopLevel.kt"),
-	UseDataClass("/cases/UseDataClass.kt"),
+	UseDataClassNegative("/cases/UseDataClassNegative.kt"),
+	UseDataClassPositive("/cases/UseDataClassPositive.kt"),
 	UnconditionalJumpStatementInLoopNegative("/cases/UnconditionalJumpStatementInLoopNegative.kt"),
 	UnconditionalJumpStatementInLoopPositive("/cases/UnconditionalJumpStatementInLoopPositive.kt"),
 	NestedClassesVisibility("/cases/NestedClassesVisibility.kt");
