+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/style/UnnecessaryInheritanceSpec.kt
@@ -15,9 +15,13 @@ class UnnecessaryInheritanceSpec : SubjectSpek<UnnecessaryInheritance>({
 		it("has unnecessary super type declarations") {
 			val findings = subject.lint("""
 				class A : Any()
-				class B : Object()
-				class C : An()""")
+				class B : Object()""")
 			assertThat(findings).hasSize(2)
+		}
+
+		it("has no unnecessary super type declarations") {
+			val findings = subject.lint("class C : An()")
+			assertThat(findings).hasSize(0)
 		}
 	}
 })
