+++ detekt-rules/src/test/resources/cases/UtilityClassesNegative.kt
@@ -2,37 +2,7 @@
 
 package cases
 
-class UtilityClassWithDefaultConstructor { // violation
-
-	companion object {
-		val C = 0
-	}
-}
-
-class UtilityClassWithPrimaryConstructor1 constructor() { // violation
-
-	companion object {
-		val C = 0
-	}
-}
-
-class UtilityClassWithPrimaryConstructor2() { // violation
-
-	companion object {
-		val C = 0
-	}
-}
-
 class UtilityClassWithPrimaryConstructorOk private constructor() {
-
-	companion object {
-		val C = 0
-	}
-}
-
-class UtilityClassWithSecondaryConstructor { // violation
-
-	constructor()
 
 	companion object {
 		val C = 0
